Use The Stubborn Network to test `sign in`
