+++ Books/BooksUITests/BooksUITests.swift
@@ -15,22 +15,24 @@ class BooksUITests: XCTestCase {
 
     override func setUp() {
         app = XCUIApplication()
-        let p = ProcessInfo()
 
-        app.launchEnvironment["TESTING"] = p.environment["TESTING"]
+        let processInfo = ProcessInfo()
+        app.launchEnvironment["STUB_PATH"] = "\(processInfo.environment["PROJECT_DIR"] ?? "")/stubs"
+        app.launchEnvironment["THE_STUBBORN_NETWORK_UI_TESTING"] = "YES"
+        app.launchEnvironment["STUB_NAME"] = self.name
 
         app.launch()
     }
 
-    func testWaveVisibleAfterSignIn() {
+    func testSignIn() {
         // given
         let usernameTextField = app.textFields["username"]
         usernameTextField.tap()
-        usernameTextField.typeText("123456789")
+        usernameTextField.typeText("577544816")
 
         let passwordTextField = app.textFields["password"]
         passwordTextField.tap()
-        passwordTextField.typeText("abcd")
+        passwordTextField.typeText("xxx")
 
         // when
         app.buttons["Sign in"].tap()
