--- LibraryCore/Sources/LibraryCore/Authentication/Authentication Manager/AuthenticationManager+Stubbed.swift
@@ -1,31 +1,0 @@
-//
-//  AuthenticationManagerUITestingStub.swift
-//  
-//
-//  Created by Martin Kim Dung-Pham on 05.07.19.
-//
-
-import Foundation
-
-extension AuthenticationManager {
-
-    /// AuthenticationManager can be stubbed for testing purposes
-    ///
-    /// - Parameter stub: the stub to set expected values on
-    ///
-    /// *Usage:*
-    /// ```
-    /// let authenticationManagerStub = AuthenticationManager.stubbed {
-    ///     $0.authenticated = true
-    ///     $0.error = nil
-    /// }
-    /// ```
-    ///
-    public static func stubbed(_ stub: ((AuthenticationManagerStub) -> Void)) -> AuthenticationManager {
-        let stubbedAuthenticationManager = AuthenticationManagerStub.init()
-
-        stub(stubbedAuthenticationManager)
-
-        return stubbedAuthenticationManager
-    }
-}
