+++ LibraryCore/Sources/LibraryCore/Extensions/ProcessInfo+Testing.swift
@@ -8,10 +8,10 @@
 import Foundation
 
 public extension ProcessInfo {
-    /// Returns `true` if the `ProcessInfo` contains a `TESTING` environment variable.
-    var testing: Bool {
+    /// Returns `true` if the `ProcessInfo` contains a `THE_STUBBORN_NETWORK_UI_TESTING` environment variable.
+    var isUITesting: Bool {
         get {
-            return environment["TESTING"] != nil
+            return environment["THE_STUBBORN_NETWORK_UI_TESTING"] != nil
         }
     }
 }
